tracker_demo/requirements.txt
-r ../requirements.txt

django-allauth==0.27.0
